Move encoding to pkg/lib/k8s
